requirements.txt
attrs==21.4.0
brother-ql==0.9.4
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
future==0.18.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packbits==0.6
Pillow==9.1.0
pyusb==1.2.1
rpimotorlib==3.1
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
